+++ udata/core/topic/forms.py
@@ -1,4 +1,5 @@
 from udata.forms import ModelForm, fields, validators
+from udata.core.spatial.forms import SpatialCoverageField
 from udata.i18n import lazy_gettext as _
 
 from .models import Topic
@@ -19,6 +20,10 @@ class TopicForm(ModelForm):
 
     datasets = fields.DatasetListField(_('Associated datasets'))
     reuses = fields.ReuseListField(_('Associated reuses'))
+
+    spatial = SpatialCoverageField(
+        _('Spatial coverage'),
+        description=_('The geographical area covered by the data.'))
 
     tags = fields.TagField(_('Tags'), [validators.DataRequired()])
     private = fields.BooleanField(_('Private'))