+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add downloads' count to organizations CSV [#2973](https://github.com/opendatateam/udata/pull/2973)
 - Add 3 new badges to the organization model : `company`, `association` and `local authority` [#2984](https://github.com/opendatateam/udata/pull/2984)
 - Prevent geozones listed ad `deleted` to be loaded [#2983](https://github.com/opendatateam/udata/pull/2983)
+- Topic: add spatial field [#2988](https://github.com/opendatateam/udata/pull/2988)
 - Topic: add last_modified field [#2987](https://github.com/opendatateam/udata/pull/2987)
 - Add stacktraces to CSV errors [#2990](https://github.com/opendatateam/udata/pull/2990)